requirements.txt
icrawler
beautifulsoup4
lxml
requests>=2.9.1
six>=1.10.0
Pillow